requirements.txt
websocket-client==1.8.0
ccxt==4.4.36
pandas==2.2.3
matplotlib==3.9.3
python-telegram-bot==21.8
python-dotenv==1.0.1
mplfinance==0.12.10b0